+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All Notable changes to `bakame/http-strucured-fields` will be documented in this
 - `InnerList::fromAssociative` to improve InnerList public API;
 - `Item` implements the `ValueAccess` interface;
 - `Item::toPair` to complement `Item::fromPair`;
+- `Item::fromDate` to improve and complete the Item Date public API;
 - `Value` internal class to improve Item public API;
 - `Token::toString` to return the string representation of the token.
 - Adding support for `MapKey` object to access container members.